Remove HttpResponseBuilder::json2()
It's not necessary to keep both json() and json2() around since the
former reduces the ownership of its parameter to a borrow only to pass
the reference to the latter. Users can instead borrow themselves when
passing an owned value: there doesn't need to be two separate functions.

This change also makes HttpResponseBuilder::json() take a value by
reference since it's cheaper to convert from an owned value to a
borrowed one rather than the other way around.